Pass enclosing instance types to `DisplayNameGenerators` (#4266)
Prior to this commit, a `DisplayNameGenerator` could access the type of 
a `@Nested` test class  as well as the enclosing class in which a
`@Nested` test class is declared, but it could not access the concrete 
runtime type of the enclosing instance for a `@Nested` test class.

When a `DisplayNameGenerator` is used to build hierarchical display
names, this could lead to confusing results or even conflicting results 
depending on the structure of the test classes used, "conflicting" in
the sense that two nested test classes may have identical display names
that do not represent the runtime structure of the test classes.

Now, each `DisplayNameGenerator` receives the list of enclosing instance
types when computing the display name of a nested test class or test 
method.

Resolves #4130.

---------

Co-authored-by: Sam Brannen <104798+sbrannen@users.noreply.github.com>