Support `@TempDir` constructor injection for Java record classes
The `@TempDir` annotation on a constructor parameter is copied to the
generated `final` instance field. This caused it top be picked up by
`TempDirectory` for instance field injection which then failed because
the field was `final`. This change now checks if a test class is an
instance of a record type and skips instance field injection. Since
records cannot have any instance fields besides the ones generated from
their constructor, there's no need to inject values into any of them.