Check for `DisplayNameGeneration` annotations on runtime enclosing types
Prior to this commit, the `DisplayNameGeneration` annotation was
discovered on the current test, a superclass, or the enclosing class 
in which a `Nested` test class is declared. However, the runtime type of
an enclosing instance is not always identical to the class in which the
`Nested` test class was declared. This caused annotations on runtime
enclosing types to be missed when checking for them in 
`IndicativeSentences` which resulted in unexpected display names. Now, 
the runtime enclosing types are checked rather than the compile-time 
ones.

Resolves #4131.

---------

Co-authored-by: Sam Brannen <104798+sbrannen@users.noreply.github.com>