Introduce @⁠SentenceFragment for IndicativeSentences display names
Prior to this commit, it was not possible to provide a custom display
name for an individual "sentence fragment" for the IndicativeSentences
DisplayNameGenerator.

Although it is possible to register a custom DisplayNameGenerator (such
as ReplaceUnderscores) to affect how a sentence fragment is generated,
you previously could not provide an entire custom sentence fragment. In
addition, attempting to use @⁠DisplayName for a sentence fragment
results in the entire sentence (that would otherwise be generated by
the IndicativeSentences DisplayNameGenerator) getting overridden by the
value of the @⁠DisplayName for the current test class or test
method.

To address that shortcoming, this commit introduces a new
@⁠SentenceFragment annotation that can be used to specify custom
sentence fragments when using the IndicativeSentences
DisplayNameGenerator.

Closes #4347
Closes #4349