Add support for Kotlin suspending functions as test/lifecycle methods (#4530)
For this to work three Kotlin dependencies are required to be on the
classpath or module path: `kotlin-stdlib`, `kotlin-reflect`, and
`kotlinx-coroutines-core`. They are likely to be already present, if
Kotlin coroutines are used in production code. The new
`KotlinReflectionUtils` checks for their availability when it needs them
but only if a method has been found that looks like a suspending
function.

Resolves #1914.