Set Java source & target compatibility to JDK 21 in Eclipse
This ensures that all code (including tests and the API tools in the
'documentation' project that make use of JDK 21 language features) is
compiled as if we were targeting JDK 21 instead of JDK 17.

Note, however, that this only affects the projects within Eclipse. The
standard Gradle build will still fail if somebody inadvertently uses
language features or APIs post JDK 17.