Ensure LauncherConfigurationParametersTests runs in all environments
Prior to this commit, the warnsOnMultiplePropertyResources() test in
LauncherConfigurationParametersTests only ran within the Gradle build
or within IntelliJ IDEA when configured with "build" as the target
build directory.

However, other environments might not use a "build" directory. For
example, Eclipse uses a "bin" directory.

This commit therefore removes the hardcoded "build/resources/test" path
from the assertion and instead determines the URL for the resource by
loading it via the ClassLoader.

This commit also refactors the assertion to make it easier to follow.