Test status quo for ConditionEvaluationResult and its factory methods
Prior to this commit, we did not have any "unit tests" for
ConditionEvaluationResult, and while working on #4698 I noticed that we
in fact have several issues in the implementation of and documentation
for ConditionEvaluationResult.

Thus, this commit introduces dedicated unit tests for the status quo,
where individual TODOs will be addressed in separate issues/commits.

See #4698
See https://github.com/junit-team/junit-framework/pull/4699#issuecomment-3018929461