Introduce support for cancelling a running execution to the Launcher API
This commit introduces `CancellationToken` that can be created and
passed to the `Launcher` as part of a `LauncherExecutionRequest`. The
`Launcher` checks whether cancellation has been requested on the token
prior to asking each test engine to execute tests. If cancellation has
been requested, the `Launcher` reports all direct children of engine
descriptors as skipped. Moreover, it passes the `CancellationToken` to
each engine so they can check and respond to cancellation as well which
will be implemented separately as indicated by TODO comments.

`EngineTestKit` also supports passing a `CancellationToken` to the
engine under test.

The documentation now contains an example for implementing a fail-fast
listener and documents the additional requirement for test engines that
wish to support cancellation.

Resolves #1880.