Allow `Stream<? extends TestTemplateInvocationContext>` return types
For consistency with `ClassTemplateInvocationContextProvider` and to
make implementations simpler, this commit changes the return type of the
`provideTestTemplateInvocationContexts(ExtensionContext)` method of the
`TestTemplateInvocationContextProvider` interface to allow more concrete
`Streams` to be returned.

This change is source compatible, meaning implementations will still
compile when declaring `Stream<TestTemplateInvocationContext>` as method
return type. To check for binary compatibility, I used Pioneer's
`RetryingTestExtension` with a 6.0.0-SNAPSHOT that included this change.
The extension kept working. Nevertheless, I included this change as
potentially breaking in the release notes.

Resolves #3577.