Quote text-based arguments in display names for parameterized tests
This commit introduces quoteTextArguments attributes in
@⁠ParameterizedClass and @⁠ParameterizedTest which default to true.

This new feature automatically encloses any CharSequence (such as a
String) in double quotes (") and any Character in single quotes (').
Special characters will be escaped within quoted text. For example,
'\n', '\r', will be escaped as "\\r" and "\\n", respectively. In
addition, any ISO control character will be represented as a question
mark (?) in the quoted text.

For example, given a String argument "line 1\nline 2", the
representation in the display name would be "\"line 1\\nline 2\""
(visually "line 1\nline 2") with the newline character escaped as
"\\n". Similarly, given a String argument "\t", the representation in
the display name would be "\"\\t\"" (visually "\t") instead of a blank
string or invisible tab character. The same applies for a character
argument '\t', whose representation in the display name would be
"'\\t'" (visually '\t').

Closes #4716