Support CSV headers with {argumentsWithNames} display name patterns
Prior to this commit, display names using CSV header names did not work
properly when using the {argumentsWithNames} display name pattern for
parameterized tests. Specifically, both the parameter name and the CSV
header name were present in the display name, leading to duplication --
for example:

[1] fruit = FRUIT = apple

This commit ensures that the parameter name is not displayed when the
user has enabled the useHeadersInDisplayName flag in @⁠CsvSource and
@⁠CsvFileSource. With this change the above display name is now
properly generated as follows (with the superfluous "fruit = " removed).

[1] FRUIT = apple

See #4716
Closes #4783