Consistently use "name = value" arguments format in parameterized tests
Prior to this commit, name-value pairs for arguments in display names
for parameterized classes and tests had different formatting/spacing
depending on the use case.

When using the `useHeadersInDisplayName` feature of @⁠CsvSource or
@⁠CsvFileSource, they took the form "name = value"; whereas, when using
the {argumentsWithNames} display name pattern, they took the form
"name=value" (without spaces).

With this commit, we now consistently use the "name = value" formatting
(with spaces).

Closes #4786