Introduce selectClasses*() variants in DiscoverySelectors
Prior to this commit, it was possible to select multiple classes by
invoking selectClass(<class>) multiple times; however, it could become
a bit verbose as in the following example.

EngineTestKit.engine("junit-jupiter")
   .selectors(
      selectClass(TestClass1A.class),
      selectClass(TestClass1B.class),
      selectClass(TestClass2A.class),
      selectClass(TestClass2B.class))

To improve the programming model for users, this commit introduces
multiple selectClasses(...) and selectClassesByName(...) variants in
DiscoverySelectors. The following example demonstrates use of the new
selectClasses(Class<?>...) varargs variant.

EngineTestKit.engine("junit-jupiter")
   .selectors(
      selectClasses(
         TestClass1A.class,
         TestClass1B.class,
         TestClass2A.class,
         TestClass2B.class))

In addition, the AdditionalDiscoverySelectors utility used by the
@⁠Suite engine no longer needs to implement its own selectClasses(...)
utilities.

Closes #4860
Closes #4863